test: centralize DOM polyfills
